# openapi.yaml
openapi: 3.1.0
info:
  title: Archer API
  version: 0.1.0
  description: API for the Archer freelance marketplace. Amounts use integer minor units; Archer does not process payments.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: Discover open jobs
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: skill, in: query, schema: { type: string } }
        - { name: workType, in: query, schema: { $ref: "#/components/schemas/WorkType" } }
        - { name: currency, in: query, schema: { $ref: "#/components/schemas/Currency" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses: { "200": { description: Paginated jobs } }
    post:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Create a draft job
      responses: { "201": { description: Job created } }
  /jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: Get a public job
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "200": { description: Job }, "404": { $ref: "#/components/responses/NotFound" } }
    patch:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Update an owned job
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "200": { description: Updated job } }
  /jobs/{jobId}/publish:
    post:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Publish a draft job
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "200": { description: Published job } }
  /jobs/{jobId}/saved:
    put:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Save a job
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "204": { description: Saved } }
    delete:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Unsave a job
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "204": { description: Removed } }
...
  /conversations:
    get:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      summary: List conversations
      responses: { "200": { description: Conversations } }
    post:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      summary: Start an eligible conversation
      responses: { "201": { description: Conversation created } }
  /conversations/{conversationId}/messages:
    get:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      summary: List messages
      parameters: [{ $ref: "#/components/parameters/ConversationId" }]
      responses: { "200": { description: Messages } }
    post:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      summary: Send a message
      parameters: [{ $ref: "#/components/parameters/ConversationId" }]
      responses: { "201": { description: Message sent } }
  /notifications:
    get:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      summary: List current user's notifications
      responses: { "200": { description: Notifications } }
  /contracts/{contractId}/reviews:
    post:
      tags: [Reviews]
      security: [{ bearerAuth: [] }]
      summary: Review the other contract party
      parameters: [{ $ref: "#/components/parameters/ContractId" }]
      responses: { "201": { description: Review created } }
  /reports:
    post:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      summary: Report content or a user
      responses: { "201": { description: Report created } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    JobId: { name: jobId, in: path, required: true, schema: { type: string } }
...
    ContractId: { name: contractId, in: path, required: true, schema: { type: string } }
...
    ConversationId: { name: conversationId, in: path, required: true, schema: { type: string } }
  schemas:
    Currency: { type: string, enum: [USD, MMK] }
    WorkType: { type: string, enum: [FIXED_PRICE, HOURLY] }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: { code: { type: string }, message: { type: string }, fields: { type: object, additionalProperties: { type: array, items: { type: string } } }, requestId: { type: string } }
  responses:
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...